+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "protoc",
+  "name": "protoc-helper",
   "version": "1.1.3",
-  "description": "A tool for converting proto buffers (.proto) files into javascript files usable in Closure Compiler with Closure Library.",
+  "description": "multilanguage  grpc protobuff generator for node. forked from node-protoc.git. ",
   "main": "index.js",
   "scripts": {
     "postinstall": "node scripts/postinstall.js",
@@ -10,16 +10,17 @@
   "bin": {
     "protoc": "./bin/protoc"
   },
-  "author": "Jeppe Rune Mortensen <jepperm@gmail.com>",
+  "author": "ji & Jeppe Rune Mortensen ",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/YePpHa/node-protoc.git"
+    "url": "git+https://github.com/ji-podhead/protoc-helper"
   },
   "dependencies": {
     "glob": "^7.2.3",
     "mkdirp": "^0.5.6",
     "node-fetch": "^3.2.10",
+        "protoc-gen-ts": "^0.8.7",
     "rimraf": "^3.0.2",
     "unzipper": "^0.10.11",
     "uuid": "^9.0.0",
